requirements.txt
anthropic==0.52.2
beautifulsoup4==4.13.4
datasets==3.5.1
duckduckgo-search==8.0.1
faiss-cpu==1.11.0
gradio==5.29.0
hf_xet==1.1.2
huggingface-hub==0.30.2
langchain==0.3.25
langchain-anthropic==0.3.15
langchain-community==0.3.24
langchain-core==0.3.64
langchain-groq==0.3.2
langchain-huggingface==0.2.0
langchain-openai==0.3.21
langchain-tavily==0.1.6
langfuse==2.60.5
langgraph==0.4.1
markdownify==1.1.0
numpy==2.2.5
openai-whisper==20240930
openpyxl==3.1.5
pandas==2.2.3
playwright==1.52.0
pyrootutils~=1.0.4
python-dotenv~=1.1.0
requests==2.32.3
tabulate==0.9.0
unstructured[pdf,docx,pptx]==0.17.2
wikipedia~=1.4.0
youtube-transcript-api==1.0.3
yt-dlp==2025.4.30
